refactor(model): model/utils -> model/util
